section 22.11

+++ admin/includes/view_all_posts.php
@@ -104,12 +104,14 @@ if(isset($_POST['checkBoxArray'])){
             echo "<td>{$post_comments}</td>";
             echo "<td width='400'>{$post_content}</td>";
             echo "<td>{$post_date}</td>";
-            echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+            echo "<td><a onclick=\" javascript: return confirm('Are you sure you want to delete?')\" href='posts.php?delete={$post_id}'>Delete</a></td>";
             echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
             echo "</tr>";
         }
 
         ?>
+
+
 
         </tbody>
     </table>
